Third_Version/requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Progress bars and utilities
tqdm>=4.65.0

# Optional for data processing
pandas>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0

# Optional for export formats
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional for dataset management
roboflow>=1.0.0

# Optional for Kaggle datasets
kaggle>=1.5.0
